Week 5 Homework
I could not get this to work and I’m not sure what I’m doing wrong. I
ultimately cannot allocate any more time to it, so I have to submit it
as it is.

I used alerts, and things worked fine. I used jQuery to just load a
different image when the page loads initially, and that worked fine. I
copied that across to use on the submit function, but that’s where
things failed.