Week 05 Homework
Added .trim( ) so that the code will still recognize the city even if
there are spaces before or after the text.

After form has been submitted, the input for the city resets - achieved
this by using .reset

If city is not recognized, an alert will pop up saying “Enter a city!”

There are comments throughout the .js file explaining what I have coded
and why.